requirements.txt
bcrypt==2.0.0
beautifulsoup4==4.3.2
cffi==1.5.2
defusedxml==0.4.1
Django==1.7
ecdsa==0.13
gunicorn==19.4.5
nose==1.3.7
oauthlib==1.0.3
Pillow==3.2.0
psycopg2==2.6.1
pycparser==2.14
pycrypto==2.6.1
PyJWT==1.4.0
python-social-auth==0.2.16
python3-openid==3.0.10
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
wheel==0.24.0
whitenoise==3.0
